# apps/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/tasks": {
      "post": {
        "operationId": "TasksController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTaskDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Task created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a new task",
        "tags": [
          "tasks"
        ]
      },
      "get": {
        "operationId": "TasksController_list",
        "parameters": [
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "description": "Stable cursor from nextCursor; empty string starts cursor pagination",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "default": 20,
              "type": "number"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "pending",
                "processing",
                "completed",
                "failed"
              ]
            }
          },
          {
            "name": "type",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "compress",
                "convert",
                "image_watermark",
                "image_id_photo",
                "pdf_merge",
                "pdf_split",
                "pdf_to_image",
                "font_convert",
                "pdf_to_text",
                "image_to_pdf",
                "pdf_rotate",
                "pdf_watermark",
                "pdf_encrypt",
                "pdf_compress",
                "pdf_metadata",
                "pdf_rearrange",
                "pdf_from_document",
                "image_generate"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Task list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "tasks": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/TaskResponseDto"
                      }
                    },
                    "total": {
                      "type": "number"
                    },
                    "nextCursor": {
                      "type": "string",
                      "nullable": true
                    }
                  },
                  "required": [
                    "tasks",
                    "total",
                    "nextCursor"
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List user tasks",
        "tags": [
          "tasks"
        ]
      }
    },
    "/tasks/image-generate/providers": {
      "get": {
        "operationId": "TasksController_listImageGenerateProviders",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Configured image generation providers",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ImageGenerateProviderDto"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List the configured image generation providers",
        "tags": [
          "tasks"
        ]
      }
    },
    "/tasks/image-generate/quota": {
      "get": {
        "operationId": "TasksController_getImageGenerateQuota",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Today image generation quota",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImageGenerateQuotaDto"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get the daily image generation quota snapshot",
        "tags": [
          "tasks"
        ]
      }
    },
    "/tasks/image-generate/presets": {
      "get": {
        "operationId": "TasksController_listImageGeneratePresets",
        "parameters": [
          {
            "name": "lang",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Enabled prompt presets in the requested language",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ImageGeneratePresetDto"
                  }
                }
              }
            }
          }
        },
        "summary": "List AI image generation prompt presets",
        "tags": [
          "tasks"
        ]
      }
    },
    "/tasks/image-generate/sessions": {
      "get": {
        "operationId": "TasksController_listImageGenerateSessions",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Sessions derived from image generate tasks, newest first",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ImageGenerateSessionDto"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List AI image generation sessions",
        "tags": [
          "tasks"
        ]
      }
    },
    "/tasks/image-generate/sessions/{sessionId}/tasks": {
      "get": {
        "operationId": "TasksController_listImageGenerateSessionTasks",
        "parameters": [
          {
            "name": "sessionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Session tasks ordered by creation time ascending",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImageGenerateSessionTasksDto"
                }
              }
            }
          },
          "400": {
            "description": "sessionId is not a valid UUID"
          },
          "401": {
            "description": "Not authenticated"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List the tasks of one image generation session",
        "tags": [
          "tasks"
        ]
      }
    },
    "/tasks/image-generate/sessions/{sessionId}": {
      "delete": {
        "operationId": "TasksController_deleteImageGenerateSession",
        "parameters": [
          {
            "name": "sessionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Session deleted"
          },
          "400": {
            "description": "sessionId is not a valid UUID"
          },
          "401": {
            "description": "Not authenticated"
          },
          "404": {
            "description": "Session not found"
          },
          "409": {
            "description": "Session has tasks still running"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete an image generation session with its files",
        "tags": [
          "tasks"
        ]
      }
    },
    "/tasks/status": {
      "get": {
        "operationId": "TasksController_getStatuses",
        "parameters": [
          {
            "name": "ids",
            "required": true,
            "in": "query",
            "description": "Comma-separated UUIDs, at most 100 unique IDs",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Task statuses",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BatchTaskStatusDto"
                  }
                }
              }
            }
          }
        },
        "summary": "Get multiple task statuses (lightweight)",
        "tags": [
          "tasks"
        ]
      }
    },
    "/tasks/{id}": {
      "get": {
        "operationId": "TasksController_getOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Task details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Access denied"
          },
          "404": {
            "description": "Task not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get task by ID",
        "tags": [
          "tasks"
        ]
      }
    },
    "/tasks/{id}/status": {
      "get": {
        "operationId": "TasksController_getStatus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Task status",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskStatusDto"
                }
              }
            }
          }
        },
        "summary": "Get task status (lightweight)",
        "tags": [
          "tasks"
        ]
      }
    },
    "/tasks/{id}/retry": {
      "post": {
        "operationId": "TasksController_retry",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "New task created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Retry a failed task",
        "tags": [
          "tasks"
        ]
      }
    },
...
  },
  "info": {
    "title": "Utils-Plane API",
    "description": "工具平台 API 文档",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "CreateTaskDto": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "compress",
              "convert",
              "image_watermark",
              "image_id_photo",
              "pdf_merge",
              "pdf_split",
              "pdf_to_image",
              "font_convert",
              "pdf_to_text",
              "image_to_pdf",
              "pdf_rotate",
              "pdf_watermark",
              "pdf_encrypt",
              "pdf_compress",
              "pdf_metadata",
              "pdf_rearrange",
              "pdf_from_document",
              "image_generate"
            ]
          },
          "inputFileIds": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          },
          "inputConfig": {
            "type": "object"
          }
        },
        "required": [
          "type",
          "inputFileIds"
        ]
      },
      "TaskResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "userId": {
            "type": "string",
            "format": "uuid"
          },
          "type": {
            "type": "string",
            "enum": [
              "compress",
              "convert",
              "image_watermark",
              "image_id_photo",
              "pdf_merge",
              "pdf_split",
              "pdf_to_image",
              "font_convert",
              "pdf_to_text",
              "image_to_pdf",
              "pdf_rotate",
              "pdf_watermark",
              "pdf_encrypt",
              "pdf_compress",
              "pdf_metadata",
              "pdf_rearrange",
              "pdf_from_document",
              "image_generate"
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "processing",
              "completed",
              "failed"
            ]
          },
          "inputFileIds": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "inputConfig": {
            "type": "object"
          },
          "outputFileId": {
            "type": "string",
            "format": "uuid"
          },
          "progress": {
            "type": "number",
            "minimum": 0,
            "maximum": 100
          },
          "errorCode": {
            "type": "string"
          },
          "errorMessage": {
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "completedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "type",
          "status",
          "inputFileIds",
          "progress",
          "createdAt"
        ]
      },
      "ImageGenerateProviderDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "来源 id，创建任务时放进 inputConfig.providerId"
          },
          "label": {
            "type": "string",
            "description": "展示给用户的来源名称"
          },
          "capabilities": {
            "type": "array",
            "description": "该来源支持的能力。generate = 文生图，edit = 图生图；缺少 edit 时前端禁用参考图上传",
            "items": {
              "type": "string",
              "enum": [
                "generate",
                "edit"
              ]
            }
          },
          "sizes": {
            "description": "该来源支持的尺寸（\"auto\" 或 \"WxH\"）。前端据此派生画面比例档位，创建任务时把选中的原始尺寸串放进 inputConfig.size",
            "items": {
              "type": "array",
              "items": {
                "type": "string"
              }
            },
            "type": "array"
          }
        },
        "required": [
          "id",
          "label",
          "capabilities",
          "sizes"
        ]
      },
      "ImageGenerateQuotaDto": {
        "type": "object",
        "properties": {
          "limit": {
            "type": "number",
            "description": "当日生图张数上限"
          },
          "used": {
            "type": "number",
            "description": "今日已用张数（失败的生成不计数）"
          },
          "remaining": {
            "type": "number",
            "description": "今日剩余张数"
          }
        },
        "required": [
          "limit",
          "used",
          "remaining"
        ]
      },
      "ImageGeneratePresetDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "模板 id（uuid）"
          },
          "title": {
            "type": "string",
            "description": "当前语言（由 lang 查询参数决定）的模板标题"
          },
          "prompt": {
            "type": "string",
            "description": "当前语言（由 lang 查询参数决定）的提示词模板正文"
          },
          "imageStorageKey": {
            "type": "string",
            "description": "示例图在 MinIO presets 桶内的对象 key；前端用 NEXT_PUBLIC_S3_PUBLIC_URL 自行拼公网 URL"
          },
          "sortOrder": {
            "type": "number",
            "description": "排序值，升序展示"
          }
        },
        "required": [
          "id",
          "title",
          "prompt",
          "sortOrder"
        ]
      },
      "ImageGenerateSessionDto": {
        "type": "object",
        "properties": {
          "sessionId": {
            "type": "string",
            "description": "会话 id（客户端生成 uuid，创建任务时放进 inputConfig.sessionId）",
            "format": "uuid"
          },
          "title": {
            "type": "string",
            "description": "会话标题：该会话第一条任务的提示词前 20 个字符"
          },
          "taskCount": {
            "type": "number",
            "description": "会话内生图任务总数"
          },
          "createdAt": {
            "type": "string",
            "description": "会话创建时间（首条任务时间，ISO 8601 UTC）"
          },
          "updatedAt": {
            "type": "string",
            "description": "会话最近活动时间（末条任务时间，ISO 8601 UTC）"
          }
        },
        "required": [
          "sessionId",
          "title",
          "taskCount",
          "createdAt",
          "updatedAt"
        ]
      },
      "ImageGenerateSessionTasksDto": {
        "type": "object",
        "properties": {
          "tasks": {
            "description": "会话内任务，按创建时间正序排列（消息流数据源）",
            "items": {
              "type": "array",
              "items": {
                "type": "taskResponseDto"
              }
            },
            "type": "array"
          },
          "total": {
            "type": "number",
            "description": "会话内任务总数（返回条数上限 200）"
          }
        },
        "required": [
          "tasks",
          "total"
        ]
      },
      "BatchTaskStatusDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "processing",
              "completed",
              "failed",
              "not_found"
            ]
          },
          "progress": {
            "type": "number",
            "minimum": 0,
            "maximum": 100
          },
          "outputFileId": {
            "type": "string",
            "format": "uuid"
          },
          "errorCode": {
            "type": "string"
          },
          "errorMessage": {
            "type": "string"
          },
          "taskId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": [
          "status",
          "progress",
          "taskId"
        ]
      },
      "TaskStatusDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "processing",
              "completed",
              "failed"
            ]
          },
          "progress": {
            "type": "number",
            "minimum": 0,
            "maximum": 100
          },
          "outputFileId": {
            "type": "string",
            "format": "uuid"
          },
          "errorCode": {
            "type": "string"
          },
          "errorMessage": {
            "type": "string"
          }
        },
        "required": [
          "status",
          "progress"
        ]
      },
...
    }
  }
}